refactor: 재사용 가능한 공통 컴포넌트는 common 폴더에 배치 / Card -> CardPreview 컴포넌트명 변경
