+++ src/components/CardExpirationSection/CardExpirationSection.tsx
@@ -1,5 +1,5 @@
 import styles from './CardExpirationSection.module.css';
-import { InputSection } from '../InputSection/InputSection';
+import { InputSection } from '../common/InputSection/InputSection';
 import { Dispatch, SetStateAction, useState } from 'react';
 import {
   validateMonthRangeError,
