+++ src/components/CardPreview/CardPreview.stories.tsx
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from '@storybook/react';
-import Card from './Card';
+import Card from './CardPreview';
 
 const meta = {
   title: 'Card',
